Auction Service: added authentication

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -4,6 +4,7 @@ using AuctionService.Repositories;
 using AutoMapper;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AuctionService.Controllers;
@@ -39,13 +40,14 @@ public class AuctionsController : ControllerBase
         return auction;
     }
     
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: add current user as seller
-        auction.Seller = "test";
-
+        
+        auction.Seller = User.Identity.Name;
+        
         _repo.AddAuction(auction);
 
         var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -60,6 +62,7 @@ public class AuctionsController : ControllerBase
             new { auction.Id }, newAuction);
     }
     
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -67,7 +70,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
         
-        // TODO: check username == seller 
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -84,6 +87,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
     }
     
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -91,7 +95,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
         
-        // TODO: check username == seller
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _repo.RemoveAuction(auction);
